+++ ft_isprint.c
@@ -5,14 +5,14 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: ghdesfos <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/09/23 14:39:46 by ghdesfos          #+#    #+#             */
-/*   Updated: 2018/09/24 15:17:54 by ghdesfos         ###   ########.fr       */
+/*   Created: 2019/06/03 18:32:48 by ghdesfos          #+#    #+#             */
+/*   Updated: 2019/06/06 13:07:30 by ghdesfos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_isprint(int c)
+int	ft_isprint(int c)
 {
-	return (c >= 32 && c <= 126);
+	return ((int)(c >= 32 && c <= 126));
 }
